requirements.txt
# Core numerical computing
numpy>=1.24.0

# Deep Learning Framework
torch>=2.7.0
torchvision>=0.22.0


# Data science libraries
scikit-learn>=1.7.0
pandas>=2.2.0
matplotlib>=3.8.0
seaborn>=0.13.2

# Utilities
tqdm>=4.66.0
PyYAML>=6.0.1

# Development tools
jupyter>=1.0.0
pytest>=8.0.0
black>=24.0.0
flake8>=7.0.0
isort>=5.13.0
mypy>=1.8.0

# Optional: Monitoring and visualization
tensorboard>=2.16.0
# plotly>=5.19.0  # For interactive plots
# wandb>=0.17.0   # For experiment tracking

# Optional: Additional utilities
# pillow>=10.3.0  # For image processing
# h5py>=3.11.0    # For HDF5 file format
# scipy>=1.13.0   # For scientific computing